chore: Move Payload collections to match named application constants
