+++ apps/civicsignalblog/src/payload/collections/Research/Authors.js
