+++ apps/civicsignalblog/src/payload/collections/Research/Pages.js
@@ -1,14 +1,14 @@
-import canRead from "../../access/applications/main";
+import canRead from "../../access/applications/research";
 import CustomPageHeader from "../../blocks/CustomPageHeader";
 import Error from "../../blocks/Error";
 import FeaturedStories from "../../blocks/FeaturedStories";
 import LongForm from "../../blocks/LongForm";
 import PageHeader from "../../blocks/PageHeader";
 import Posts from "../../blocks/Posts";
-import pages from "../../fields/pages";
+import pages from "../../utils/createPagesCollection";
 
 const Pages = pages({
-  pageSlug: "civicsignal-pages",
+  pageSlug: "pages",
   label: "Pages",
   group: "Publication",
   defaultColumns: ["fullTitle", "updatedAt"],
@@ -22,6 +22,9 @@ const Pages = pages({
   ],
   access: {
     read: canRead,
+  },
+  adminOptions: {
+    description: "Research",
   },
 });
